requirements.txt
pytest==7.4.0
requests==2.31.0
jsonschema==4.21.1
curlify==2.2.1
selene~=2.0.0rc4
allure-pytest==2.13.2
